[#6] Refactor NNLoader to host a SuggestionsEngine
