+++ src/main/resources/public/history.html
@@ -40,11 +40,7 @@
 
     <div id="navbar"></div>
 
-    <div class="jumbotron" id="issues_container">
-        <div class="jumboback">
-        <center><h4><b>Top Issues</b></h4></center>
-        <div id="issues"></div>
-        </div></div>  <!-- /container -->
+    <div id="top_jumbotron"></div>
 
     <!-- Main component for a primary marketing message or call to action -->
     <div class="jumbotron">
